Refactor - country, member package 생성
