+++ src/main/java/world/podo/emergency/application/CountrySynchronizationApplicationService.java
@@ -3,9 +3,9 @@ package world.podo.emergency.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import world.podo.emergency.domain.Country;
-import world.podo.emergency.domain.CountryFetchService;
-import world.podo.emergency.domain.CountrySynchronizationService;
+import world.podo.emergency.domain.country.Country;
+import world.podo.emergency.domain.country.CountryFetchService;
+import world.podo.emergency.domain.country.CountrySynchronizationService;
 
 import java.util.List;
 import java.util.stream.Collectors;
