+++ src/main/java/world/podo/emergency/application/MemberApplicationService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
-import world.podo.emergency.domain.Member;
-import world.podo.emergency.domain.MemberNotFoundException;
-import world.podo.emergency.domain.MemberRepository;
+import world.podo.emergency.domain.member.Member;
+import world.podo.emergency.domain.member.MemberNotFoundException;
+import world.podo.emergency.domain.member.MemberRepository;
 
 @Service
 @RequiredArgsConstructor
