--- src/main/java/world/podo/emergency/domain/ContactRepository.java
@@ -1,6 +1,0 @@
-package world.podo.emergency.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ContactRepository extends JpaRepository<Contact, Long> {
-}
