--- src/main/java/world/podo/emergency/domain/MemberNotFoundException.java
@@ -1,4 +1,0 @@
-package world.podo.emergency.domain;
-
-public class MemberNotFoundException extends RuntimeException {
-}
