+++ src/main/java/world/podo/emergency/domain/country/ContactFactory.java
@@ -1,4 +1,4 @@
-package world.podo.emergency.domain;
+package world.podo.emergency.domain.country;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -6,10 +6,10 @@ import org.springframework.util.Assert;
 
 @Component
 @RequiredArgsConstructor
-public class ContactFactory {
+class ContactFactory {
     private final ContactRepository contactRepository;
 
-    public Contact create(Country country) {
+    Contact create(Country country) {
         Assert.notNull(country, "'country' must not be null");
 
         return new Contact(
