+++ src/main/java/world/podo/emergency/domain/country/ContactRepository.java
@@ -1,0 +1,6 @@
+package world.podo.emergency.domain.country;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+interface ContactRepository extends JpaRepository<Contact, Long> {
+}
