+++ src/main/java/world/podo/emergency/domain/country/CountryFetchService.java
@@ -1,4 +1,4 @@
-package world.podo.emergency.domain;
+package world.podo.emergency.domain.country;
 
 import java.util.List;
