+++ src/main/java/world/podo/emergency/domain/country/CountryNotFoundException.java
@@ -1,4 +1,4 @@
-package world.podo.emergency.domain;
+package world.podo.emergency.domain.country;
 
 public class CountryNotFoundException extends RuntimeException {
     private static final long serialVersionUID = -4369305816206888932L;
