+++ src/main/java/world/podo/emergency/domain/country/Notice.java
@@ -1,4 +1,4 @@
-package world.podo.emergency.domain;
+package world.podo.emergency.domain.country;
 
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
