+++ src/main/java/world/podo/emergency/domain/country/NoticeFetchService.java
@@ -1,4 +1,4 @@
-package world.podo.emergency.domain;
+package world.podo.emergency.domain.country;
 
 import java.util.List;
