+++ src/main/java/world/podo/emergency/domain/member/MemberNotFoundException.java
@@ -1,0 +1,5 @@
+package world.podo.emergency.domain.member;
+
+public class MemberNotFoundException extends RuntimeException {
+    private static final long serialVersionUID = -8476312253058941214L;
+}
