+++ src/main/java/world/podo/emergency/infrastructure/public_api/PublicApiContactFetchService.java
@@ -9,8 +9,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
-import world.podo.emergency.domain.ContactFetchService;
-import world.podo.emergency.domain.ContactFetchValue;
+import world.podo.emergency.domain.country.ContactFetchService;
+import world.podo.emergency.domain.country.ContactFetchValue;
 
 import java.net.URI;
 import java.util.List;
