+++ src/main/java/world/podo/emergency/ui/batch/ContactSynchronizationScheduler.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import world.podo.emergency.application.CountrySynchronizationApplicationService;
-import world.podo.emergency.domain.Country;
+import world.podo.emergency.domain.country.Country;
 
 import java.util.List;
