+++ src/main/java/world/podo/emergency/ui/web/ApiControllerAdvice.java
@@ -7,7 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import world.podo.emergency.domain.MemberNotFoundException;
+import world.podo.emergency.domain.member.MemberNotFoundException;
 import world.podo.emergency.infrastructure.public_api.PublicApiFailedException;
 import world.podo.emergency.infrastructure.spring.security.PodoAuthenticationToken;
